requirements.txt
anyio==3.3.4
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
outcome==1.1.0
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==21.0.0
requests==2.26.0
selenium
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
SQLAlchemy==1.4.26
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.4
starlette==0.16.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
webdriver-manager==3.5.2
wsproto==1.0.0
alembic==1.7.1
asyncpg==0.24.0
psycopg2-binary==2.9.1
python-dotenv
python-logstash-async==2.3.0    
starlette==0.16.0   
elastic-apm==6.7.2    